+++ packages/ranger-hooks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ranger-theme/hooks
 
+## 0.9.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @ranger-theme/utils@0.11.0
+
 ## 0.9.1
 
 ### Patch Changes